7.vector-db/requirements.txt
chromadb>=0.4.0
sentence-transformers>=2.2.0
requests>=2.28.0
python-dotenv>=1.0.0
tqdm>=4.65.0